+++ Classes/Transformers/SFBillSearchCellTransformer.m
@@ -29,7 +29,7 @@ static TTTOrdinalNumberFormatter *ordinalNumberFormatter;
     NSInteger introYear = [bill.introducedOn dateComponents].year;
     NSString *humanSession = [NSString stringWithFormat:@"%@ Congress", [ordinalNumberFormatter stringFromNumber:bill.congress]];
     cellData.detailTextLabelString = humanSession;
-    cellData.tertiaryTextLabelString = [NSString stringWithFormat:@"%d", introYear];
+    cellData.tertiaryTextLabelString = [NSString stringWithFormat:@"%ld", (long)introYear];
 
     return cellData;
 }
